refactor: combine `keys` and `key_types`
