+++ securedrop-protocol/protocol-minimal/src/lib.rs
@@ -4,7 +4,6 @@ extern crate alloc;
 pub mod api;
 mod ciphertext;
 mod constants;
-mod key_types;
 pub mod keys;
 pub mod messages;
 pub mod primitives;
@@ -19,7 +18,7 @@ pub use constants::{LEN_DH_ITEM, LEN_MLKEM_ENCAPS_KEY, LEN_XWING_ENCAPS_KEY};
 
 pub use ciphertext::{CombinedCiphertext, Envelope, FetchResponse, Plaintext};
 
-pub use key_types::{
+pub use keys::{
     DhAkemKeyPair, DhFetchKeyPair, Enrollment, KeyBundlePublic, KeyPair, MlKem768KeyPair,
     SessionStorage, SignedKeyBundlePublic, SignedLongtermPubKeyBytes, SigningKeyPair, XWingKeyPair,
 };
@@ -29,7 +28,7 @@ pub use traits::{Enrollable, JournalistPublic, UserPublic, UserSecret};
 pub use journalist::{Journalist, JournalistPublicView};
 pub use source::{Source, SourcePublicView};
 
-pub(crate) use key_types::MessageKeyBundle;
+pub(crate) use keys::MessageKeyBundle;
 pub(crate) use traits::private;
 
 // Primitives for signing
